# backend/api-docs/api-docs.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "https://localhost:8443",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/apartments/{id}": {
      "get": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "getApartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "updateApartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "createApartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "deleteApartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reviews/post/{id}": {
      "post": {
        "tags": [
          "review-rest-controller"
        ],
        "operationId": "postReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{id}/image": {
      "get": {
        "tags": [
          "user-rest-controller"
        ],
        "operationId": "downloadImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reviews/{id}": {
      "get": {
        "tags": [
          "review-rest-controller"
        ],
        "operationId": "getReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/loadMore/{start}/{end}": {
      "get": {
        "tags": [
          "review-rest-controller"
        ],
        "operationId": "loadMoreReviews",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "end",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/apartments/{id}/info": {
      "get": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "apartmentInformation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Apartment"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apartments/{id}/images": {
      "get": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "downloadImage_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apartments/manager/loadMore/{start}/{end}": {
      "get": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "loadMoreApartmentsManagerView",
        "parameters": [
          {
            "name": "start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "end",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Apartment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apartments/loadMore": {
      "get": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "loadMoreApartments",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "numPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 6
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelApartmentDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apartments/": {
      "get": {
        "tags": [
          "apartment-rest-controller"
        ],
        "operationId": "getApartments",
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentDTO"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApartmentBasicDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "format": "float"
          },
          "imagePath": {
            "type": "string"
          },
          "image": {
            "type": "boolean"
          }
        }
      },
      "ReservationBasicDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "numPeople": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ReviewBasicDTO": {
        "type": "object",
        "properties": {
          "reviewId": {
            "type": "integer",
            "format": "int64"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string"
          }
        }
      },
...
      "ApartmentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "format": "float"
          },
          "imagePath": {
            "type": "string"
          },
          "image": {
            "type": "boolean"
          },
          "manager": {
            "$ref": "#/components/schemas/UserBasicDTO"
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomBasicDTO"
            }
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationBasicDTO"
            }
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewBasicDTO"
            }
          }
        }
      },
      "RoomBasicDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "maxClients": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "UserBasicDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "type": "boolean"
          },
          "imagePath": {
            "type": "string"
          },
          "rols": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nick": {
            "type": "string"
          }
        }
      },
      "Apartment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "format": "float"
          },
          "imageFile": {
            "type": "object",
            "properties": {
              "binaryStream": {
                "type": "object"
              }
            }
          },
          "image": {
            "type": "boolean"
          },
          "imagePath": {
            "type": "string"
          },
          "manager": {
            "$ref": "#/components/schemas/UserE"
          },
          "numRooms": {
            "type": "integer",
            "format": "int32"
          },
          "averageRating": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "apartment": {
            "$ref": "#/components/schemas/Apartment"
          },
          "user": {
            "$ref": "#/components/schemas/UserE"
          }
        }
      },
      "UserE": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "imageFile": {
            "type": "object",
            "properties": {
              "binaryStream": {
                "type": "object"
              }
            }
          },
          "image": {
            "type": "boolean"
          },
          "imagePath": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "validated": {
            "type": "boolean"
          },
          "rejected": {
            "type": "boolean"
          },
          "rols": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nick": {
            "type": "string"
          },
          "pass": {
            "type": "string"
          },
          "apartment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Apartment"
            }
          }
        }
      },
      "ReviewDTO": {
        "type": "object",
        "properties": {
          "reviewId": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserBasicDTO"
          },
          "apartment": {
            "$ref": "#/components/schemas/ApartmentBasicDTO"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelApartmentDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApartmentDTO"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      }
    }
  }
}